Expand available map pan to account for future DLC
The changes to camera_limits affect the World Map. [4] is the highest possible zoom (largest scale), [7] is the lowest zoom (smallest scale). SCS applies a slight progression to those limits, but it's somewhat inconsistent.

The changes to ui_map_camera affect the other maps, e.g. in the Job Chooser. SCS tends to set these to about the same value as for the World Map at lowest zoom. I chose a slightly smaller value, which appears to work better with the expanded map pan.